backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
smbprotocol
Pillow
httpx
python-dotenv
python-multipart
torch
torchvision
nudenet
ultralytics
transformers
sentencepiece
facenet-pytorch
scipy
reverse_geocode
watchfiles
imagehash
